backend/requirements.txt
numpy>=1.24.0
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
scipy>=1.7.0
threadpoolctl>=2.2.0
joblib>=1.1.0
httpx>=0.23.0
python-jose[cryptography]>=3.3.0
pyjwt>=2.3.0
python-multipart>=0.0.5
python-dateutil>=2.8.2
Pillow>=8.4.0
openai>=0.27.0
elevenlabs>=0.2.12
python-magic>=0.4.24
python-magic-bin>=0.4.14; sys_platform == 'win32'
fastapi>=0.104.0
uvicorn>=0.24.0
groq>=0.4.0
